Add module-info.java for standard Kotlin libraries
Using the new multi-release jar feature, store compiled
module-info.class files into META-INF/versions/9 instead of the artifact
root to allow the artifacts be usable on Java versions less than 9.

For kotlin-reflect, the standard way of loading resources with built-ins
metadata from the current class loader no longer works in the modular
mode on Java 9 because the current class loader is for module
'kotlin.reflect', but the metadata is located in the module
'kotlin.stdlib'. On Java 9, we now use the class loader of
`kotlin.stdlib` to load these resources.

Temporarily switch to a yet unreleased Shadow plugin version
2.0.2-SNAPSHOT, which supports relocation of module-info files

 #KT-21266 Fixed